+++ packages/lockfile-lint-api/__tests__/parseYarnLockfile.test.js
@@ -24,6 +24,26 @@ describe('ParseLockfile Yarn', () => {
     )
   })
 
+  it('providing empty content for lockfileText throws an error', () => {
+    const parser = new ParseLockfile({lockfileText: '\n\n', lockfileType: 'yarn'})
+    expect(() => {
+      parser.parseSync()
+    }).toThrowError('Unable to parse yarn lockfile ""')
+  })
+
+  it('providing invalid content for lockfileText throws an error', () => {
+    const parser = new ParseLockfile({lockfileText: '{'.repeat(9000), lockfileType: 'yarn'})
+    expect(() => {
+      parser.parseSync()
+    }).toThrowError('Unable to parse yarn lockfile ""')
+  })
+  it('providing garbled content for lockfileText throws an error', () => {
+    const parser = new ParseLockfile({lockfileText: `# yarn lockfile v1`, lockfileType: 'yarn'})
+    expect(() => {
+      parser.parseSync()
+    }).toThrowError('Unable to parse yarn lockfile ""')
+  })
+
   it('parsing a yarn lockfile with invalid content throws an error', () => {
     const mockYarnLockfilePath = path.join(__dirname, './__fixtures__/bad-yarn.lock')
     const options = {