+++ packages/lockfile-lint-api/index.js
@@ -5,6 +5,7 @@ const ValidateHttps = require('./src/validators/ValidateHttps')
 const ValidatePackageNames = require('./src/validators/ValidatePackageNames')
 const ValidateScheme = require('./src/validators/ValidateScheme')
 const ValidateUrl = require('./src/validators/ValidateUrl')
+const ValidateIntegrity = require('./src/validators/ValidateIntegrity')
 const ParseLockfile = require('./src/ParseLockfile')
 
 module.exports = {
@@ -13,5 +14,6 @@ module.exports = {
   ValidateHttps,
   ValidatePackageNames,
   ValidateScheme,
-  ValidateUrl
+  ValidateUrl,
+  ValidateIntegrity
 }